WIP: serialization/deserialization working
